fix: add unauthorized error handling for accessToken in admin actions

+++ pod-pirate-frontend/actions/admin/podcasts.ts
@@ -5,6 +5,11 @@ import { auth } from "@/lib/utils";
 export async function handleAdminDeletePodcast(podcastId: string) {
   const session = await auth();
   const token = session?.user.accessToken;
+
+  if (!token) {
+    throw new Error("Unauthorized");
+  }
+
   try {
     const res = await fetch(`${process.env.NEXT_BACKEND_URL}/api/admin/podcasts/${podcastId}`, {
       method: "DELETE",
@@ -34,6 +39,11 @@ export async function handleAdminDeletePodcast(podcastId: string) {
 export async function handleFlagPodcast(podcastId: string) {
   const session = await auth();
   const token = session?.user.accessToken;
+
+  if (!token) {
+    throw new Error("Unauthorized");
+  }
+
   try {
     const res = await fetch(
       `${process.env.NEXT_BACKEND_URL}/api/admin/podcasts/${podcastId}/flag`,
@@ -66,6 +76,9 @@ export async function handleFlagPodcast(podcastId: string) {
 export async function handleUnflagPodcast(podcastId: string) {
   const session = await auth();
   const token = session?.user.accessToken;
+  if (!token) {
+    throw new Error("Unauthorized");
+  }
   try {
     const res = await fetch(
       `${process.env.NEXT_BACKEND_URL}/api/admin/podcasts/${podcastId}/unflag`,
